update Identity class composite key

+++ src/main/java/app/repository/CommentsRepository.java
@@ -3,5 +3,5 @@ package app.repository;
 import app.model.Comments;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentsRepository extends JpaRepository<Comments,Long> {
+public interface CommentsRepository extends JpaRepository<Comments,Integer> {
 }
